+++ examples/fullstack/contact-forms/src/pages/app/contact-forms/index.tsx
@@ -7,7 +7,7 @@ import {
   Stack,
   Text,
 } from '@chakra-ui/react'
-import { useE2ESDKClientKeys } from '@e2esdk/react'
+import { useE2ESDKClientKeys } from '@socialgouv/e2esdk-react'
 import { NoSSR } from 'components/NoSSR'
 import type { NextPage } from 'next'
 import NextLink from 'next/link'