+++ examples/with-vue/src/App.vue
@@ -1,11 +1,17 @@
 <script setup lang="ts">
-import '@e2esdk/devtools';
-import HelloWorld from './components/HelloWorld.vue';
+import '@socialgouv/e2esdk-devtools'
+import HelloWorld from './components/HelloWorld.vue'
 </script>
 
 <template>
   <header>
-    <img alt="Vue logo" class="logo" src="@/assets/logo.svg" width="125" height="125" />
+    <img
+      alt="Vue logo"
+      class="logo"
+      src="@/assets/logo.svg"
+      width="125"
+      height="125"
+    />
 
     <div class="wrapper">
       <HelloWorld msg="You did it!" />