+++ packages/devtools/src/components/SodiumState.tsx
@@ -1,5 +1,5 @@
 import { Badge, BadgeProps, Icon, Text } from '@chakra-ui/react'
-import { useE2ESDKClient } from '@e2esdk/react'
+import { useE2ESDKClient } from '@socialgouv/e2esdk-react'
 import React from 'react'
 import { FiCheckCircle, FiLoader } from 'react-icons/fi'