+++ packages/server/src/routes/participants.ts
@@ -4,7 +4,7 @@ import {
   GetParticipantsResponseBody,
   publicKeyAuthHeaders,
   PublicKeyAuthHeaders,
-} from '@e2esdk/api'
+} from '@socialgouv/e2esdk-api'
 import { z } from 'zod'
 import { zodToJsonSchema } from 'zod-to-json-schema'
 import { getNamePayloadParticipantsWithPermissions } from '../database/models/participants.js'