+++ packages/server/src/routes/permissions.ts
@@ -6,7 +6,7 @@ import {
   PostPermissionRequestBody,
   PublicKeyAuthHeaders,
   publicKeyAuthHeaders,
-} from '@e2esdk/api'
+} from '@socialgouv/e2esdk-api'
 import { z } from 'zod'
 import { zodToJsonSchema } from 'zod-to-json-schema'
 import {